Use QrCodeScanUtils to abstract barcode scanning implementation
